Fix request body for UpdateDrive
The UpdateDrive request does not require the 'name' property. It is only
required for creating a drive. We fixed this by introducing a new type
'driveUpdate' with 'name' as an optional property and deriving the
'drive' type from it with 'name' as a required property.

This is needed because the latest release of the code generator for go
now does some basic validation for required attributes.